upgrade farao dependencies version and farao-dichotomy version (#164)
* upgrade powsybl and farao core version

* upgarde farao-dichotomy version and rao runner version

+++ cse-cc-import-runner-app/src/main/java/com/farao_community/farao/cse/import_runner/app/dichotomy/ZonalScalableProvider.java
@@ -73,13 +73,13 @@ public class ZonalScalableProvider {
     }
 
     private static Scalable getStackedScalable(CseCountry cseCountry, Scalable scalable, Network network, double sum) {
-        List<Float> percentageList = new ArrayList<>();
+        List<Double> percentageList = new ArrayList<>();
         List<Scalable> scalableList = new ArrayList<>();
 
         network.getLoadStream()
             .filter(load -> isLoadCorrespondingToTheCountry(load, cseCountry))
             .forEach(load -> {
-                percentageList.add((float) (load.getP0() / sum) * 100);
+                percentageList.add((load.getP0() / sum) * 100);
                 scalableList.add(Scalable.onLoad(load.getId()));
             });
